[consensus][proxy][1/n] implement proxy routing logic
This patch implements a RoutingTable class to control the routes of
messages being proxied across a replica set topology. This routing table
is actually more of a routing tree, since there is a single-parent
relationship among nodes represented in the routing table, and so there
is only one possible route to each node at any given instance of the
configuration. That said, most people have some context for what a
routing table is used for, so that's why I chose that name. See the
comment at the top of routing.h for details on the assumptions made and
the behavior of the RoutingTable class.

This patch also implements a wrapper class called DurableRoutingTable
that persists the relevant routing metadata to disk.

The proxy routing topology is specified in a new protobuf called
ProxyTopologyPB, which is separate from RaftConfigPB so that the
topology can be changed independently of Raft config change. Such an RPC
API appears later in this patch series.

Included is a simple unit test of the next-hop routing behavior for
RoutingTable. The DurableRoutingTable could use unit tests, but for now
has integration test coverage in a later patch in this series.

Some TODOs:

* Could use some lower level test coverage for DRT including a
  persistence test and coverage for Incomplete status handling.

Change-Id: I44e9ccfe4ca97ac24c30b7163778840e1d4b823e